requirements.txt
aiofiles==23.2.1
aiogram==2.25.1
aiohttp==3.8.5
aiosignal==1.3.1
anyio==3.7.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
beautifulsoup4==4.12.2
bidict==0.22.1
bs4==0.0.1
certifi==2023.5.7
cffi==1.16.0
charset-normalizer==3.2.0
click==8.1.6
cryptography==41.0.4
defusedxml==0.7.1
Django==4.2.1
django-allauth==0.57.0
django-jazzmin==2.6.0
djangorestframework==3.14.0
drf-yasg==1.21.7
exceptiongroup==1.1.2
fastapi==0.101.0
fastapi-socketio==0.0.10
frozenlist==1.4.0
gunicorn==20.1.0
h11==0.14.0
htmx==0.0.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
idna==3.4
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.3
magic-filter==1.0.11
markdown2==2.4.10
MarkupSafe==2.1.3
multidict==6.0.4
nicegui==1.3.9
oauthlib==3.2.2
orjson==3.9.4
packaging==23.2
Pillow==9.5.0
pscript==0.7.7
psycopg2-binary==2.9.9
pycparser==2.21
pydantic==1.10.12
Pygments==2.16.1
PyJWT==2.8.0
python-dotenv==1.0.0
python-engineio==4.5.1
python-multipart==0.0.6
python-socketio==5.8.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
slugify==0.0.1
sniffio==1.3.0
soupsieve==2.4.1
sqlparse==0.4.4
starlette==0.27.0
transliterate==1.10.2
typing_extensions==4.6.2
uritemplate==4.1.1
urllib3==2.0.3
uvicorn==0.22.0
uvloop==0.17.0
vbuild==0.8.2
watchfiles==0.19.0
websockets==11.0.3
yarl==1.9.2
